+++ consumer-template/modules/stack/identity.tf
@@ -1,5 +1,5 @@
 module "identity" {
-  source = "git::https://github.com/sheyaln/sabokit.git//platform/identity/terraform?ref=v2.7.2"
+  source = "git::https://github.com/sheyaln/sabokit.git//platform/identity/terraform?ref=v2.8.0"
 
   gateway_domain = module.base.domains.gateway_domain
   base_domain    = module.base.domains.base_domain